add condition mechanism

+++ src/types/task_id.rs
@@ -1,1 +1,1 @@
-pub type TaskId = u32;
+pub type TaskId = usize;
